+++ lib/karafka/web/ui/pro/controllers/errors.rb
@@ -18,18 +18,45 @@ module Karafka
         module Controllers
           # Errors details controller
           class Errors < Ui::Controllers::Base
-            # @param partition_id [Integer] id of the partition of errors we are interested in
-            def index(partition_id)
-              errors_topic = ::Karafka::Web.config.topics.errors
-              @partition_id = partition_id
-              @previous_page, @error_messages, @next_page, @partitions_count = \
-                Models::Message.page(
-                  errors_topic,
-                  @partition_id,
-                  @params.current_page
-                )
+            include Ui::Lib::Paginations
 
+            # Lists all the errors from all the partitions
+            def index
+              @topic_id = errors_topic
+              @partitions_count = Models::ClusterInfo.partitions_count(errors_topic)
+
+              @active_partitions, materialized_page, @limited = Paginators::Partitions.call(
+                @partitions_count, @params.current_page
+              )
+
+              @error_messages, next_page = Models::Message.topic_page(
+                errors_topic, @active_partitions, materialized_page
+              )
+
+              paginate(@params.current_page, next_page)
+
+              respond
+            end
+
+            # @param partition_id [Integer] id of the partition of errors we are interested in
+            def partition(partition_id)
+              @partition_id = partition_id
               @watermark_offsets = Ui::Models::WatermarkOffsets.find(errors_topic, @partition_id)
+              @partitions_count = Models::ClusterInfo.partitions_count(errors_topic)
+
+              previous_offset, @error_messages, next_offset = Models::Message.offset_page(
+                errors_topic,
+                @partition_id,
+                @params.current_offset,
+                @watermark_offsets
+              )
+
+              paginate(
+                previous_offset,
+                @params.current_offset,
+                next_offset,
+                @error_messages.map(&:offset)
+              )
 
               respond
             end
@@ -39,7 +66,6 @@ module Karafka
             # @param partition_id [Integer]
             # @param offset [Integer]
             def show(partition_id, offset)
-              errors_topic = ::Karafka::Web.config.topics.errors
               @partition_id = partition_id
               @offset = offset
               @error_message = Models::Message.find(
@@ -49,6 +75,13 @@ module Karafka
               )
 
               respond
+            end
+
+            private
+
+            # @return [String] errors topic
+            def errors_topic
+              ::Karafka::Web.config.topics.errors
             end
           end
         end